Add support for Cassandra 1.2
To use Cassandra 1.2, the Cassandra version must specified for the require of
the cassandra-cql gem. Implement support for Cassandra 1.2 using the same
pattern that is used in the cassandra-cql gem (i.e. a module method stored in
a file named with the Cassandra version).

Bump the required cassandra-cql version to 1.2.